requirements.txt
langchain
langchain_core
langchain_google_genai
langchain_community
line-bot-sdk==3.5.0
fastapi
google.generativeai
pydantic
tiktoken
Pillow